fix(seo): fold sitemap into api/config.ts, stay under Hobby 12-function cap
Standalone api/sitemap.xml.ts pushed the deployment to 13 serverless
functions, over Vercel Hobby's limit. Both are public read-only GETs, so
sitemap generation now lives behind api/config.ts?format=sitemap instead.
